--- a/_inc/client/rest-api/index.js
+++ b/_inc/client/rest-api/index.js
@@ -122,6 +122,14 @@ const restApi = {
 		}
 	} )
 		.then( checkStatus ).then( response => response.json() ),
+	getStats: () => fetch( `${ window.Initial_State.WP_API_root }jetpack/v4/module/stats/get`, {
+		credentials: 'same-origin',
+		headers: {
+			'X-WP-Nonce': window.Initial_State.WP_API_nonce,
+			'Content-type': 'application/json'
+		}
+	} )
+		.then( checkStatus ).then( response => response.json() ),
 	getPluginUpdates: () => fetch( `${ window.Initial_State.WP_API_root }jetpack/v4/updates/plugins`, {
 		credentials: 'same-origin',
 		headers: {