--- a/_inc/client/state/at-a-glance/actions.js
+++ b/_inc/client/state/at-a-glance/actions.js
@@ -4,6 +4,9 @@
 import restApi from 'rest-api';
 import {
 	STATS_SWITCH_TAB,
+	STATS_FETCH,
+	STATS_FETCH_FAIL,
+	STATS_FETCH_SUCCESS,
 	AKISMET_DATA_FETCH,
 	AKISMET_DATA_FETCH_FAIL,
 	AKISMET_DATA_FETCH_SUCCESS,
@@ -26,6 +29,25 @@ export const statsSwitchTab = ( tab ) => {
 		dispatch( {
 			type: STATS_SWITCH_TAB,
 			activeStatsTab: tab
+		} );
+	}
+}
+
+export const fetchStats = () => {
+	return ( dispatch ) => {
+		dispatch( {
+			type: STATS_FETCH
+		} );
+		return restApi.getStats().then( stats => {
+			dispatch( {
+				type: STATS_FETCH_SUCCESS,
+				stats: stats
+			} );
+		} )['catch']( error => {
+			dispatch( {
+				type: STATS_FETCH_FAIL,
+				error: error
+			} );
 		} );
 	}
 }