A branch where I'm experimenting with performance issues with Jetpack React.

Baseline: 

Loading the Jetpack dashboard takes a pretty long time. On my dev environment, Jetpack dashboard transfers 3.6MB of data, issues 82 DB requests, 125 individual asset requests, 6 back-end API requests to WPCOM, takes ~6-8 seconds of back-end time and and takes 20 - 25 seconds to load on a moderate broadband connection. These stats don't even include back-end time spent within individual API requests to the site.

Things in this branch:
- Change timestamp to version number for admin.js and other dashboard assets. Developers who want these assets refreshed can just disable caching. Saves: ~3mb per page load.
- Allow storing NULL response for post_by_email_address, so that we don't try to retrieve this value from WPCOM on every page load. Saves: 1 WPCOM request on every dashboard load (for users without post by email address).
- Cache the output of `Jetpack::get_connected_user_data( $user_id )` for 1 day (saves 1 WPCOM request for every dashboard load)
- In-progress: Load stats via an API call rather than inline during page load. Despite caching of 5 minutes for stats, this basically incurs 4-5 additional WPCOM requests on every load of the Jetpack dashboard, since I'm guessing most times people come back to the dashboard after more than 5 minutes has passed.

WIP results:

| | Before  | After |
| --------- | --------| ----- |
| Remote requests  | 2-6 | 0 - 6 (mostly 0!) |
| Download size | 3.6 MB | 0.424 MB |
| Back-end time | 8s | 0.6s |

The upside of all this: For regular users, this should see the load time of the dashboard drop from over 10 seconds to 1-2 seconds, particularly if their cache is primed.