@gravityrail 2016-07-20
@eliorivero - as discussed, could you please finish the work I started in https://github.com/Automattic/jetpack/pull/4469/commits/9b42d1adc27ecd7a6f6e80c7aa40f4f7ddb037ae to fetch stats from the API after page load? This will save 4 calls to WPCOM from most requests.

Separately, we should look at making fewer requests during that call, either by combining the 4 requests into 1 or by only fetching the data that's being actively viewed (e.g. just the daily, monthly or yearly stats).