@eliorivero 2016-07-21
@gravityrail I've updated the PR with the commit to make Stats data load after page load. It will be cached once it's correctly fetch in the JS object we use to store initial data so when user switches tabs and backs to view Stats they don't have to be fetch again.