@gravityrail 2016-07-21
@eliorivero works great! A couple of things:

- probably should have a spinner state (didn't see one) while it's loading
- might be worth setting the height of the stats area while it's loading so the page layout doesn't jump downwards once it's loaded

What do you think? Should we merge this and address those in a separate PR?