@eliorivero 2016-07-21
@gravityrail we need to introduce skeleton loading in many views that load remote data so we'll surely do that in a separate PR. When the skeleton loading for Stats is introduced, it will take the height of the stats area so that will be addressed too.